requirements.txt
tensorflow-cpu==2.10.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
Flask==2.2.5
Werkzeug==2.2.3
gunicorn==20.1.0
